Added Rhino6 compatable version
The grasshopper example file version from rhino 5 had a bug creating opposite ramp angle when used in rhino6. The result would be wrong ramp angles in hard materials like aluminium (angle of 78 degrees instead of 12). A small fix was needed in the "ramp angle" component inside  the "post processor" component. The problem was the inputs and outputs of the "Rigth Trigonometery" component had switched between grasshopper for Rhino5 and Rhino6.