Improvements for drilling + clean up file
I added an option to use autosort with drilling, I'll add this to other operations as well in the future. I also dug into the postprocessor to check for speed improvements. I tweaked the file so that trochoidal milling doesn't add as much calculation time to the component when not activated.
And then I also removed the todolist from the file, since those have been transferred to Github issues.