New version solving endless curves in curves + postprocessor update
As discussed in issue #50  this version includes endless curves inside curves calculcation.  This has not been solved for pockets yet, but I assume the same solution will be easy to implement. Also includes a bug fix for the postprocessor. After the last changes it would not create toolpaths with ramping. Now this works again.